Added function to compute AABB vertices

+++ src/RcsCore/Rcs_utilsCPP.cpp
@@ -309,7 +309,7 @@ std::string String_concatenate(std::vector<std::string>& words, std::string sep)
 {
   std::string res;
 
-  for (int i = 0; i < words.size(); ++i)
+  for (size_t i = 0; i < words.size(); ++i)
   {
     res += words[i];
